refactor(tracking): move runs/judgement_log/trackers out of core
Observability, not ER modelling -- so it belongs beside `core`, not in it.
`git mv` throughout, so rename detection carries the history.

Membership is the approved plan's (tools/refactor_target_packages.json):
  core/runs.py            -> tracking/runs.py            (601)
  core/judgement_log.py   -> tracking/judgement_log.py   (279)
  core/trackers/          -> tracking/trackers/          (900, 4 files)
  clients/tracking.py     -> tracking/factories.py       (156)

`clients/tracking.py` is the mapping's fourth `-> tracking` entry
(refactor_target_packages.json:85). It exists, so it moves; that also
collapses the `clients <-> tracking` mutual pair, since `wandb_tracker`
imported the factory at toplevel and the factory imported `TrackioTracker`
back.

Every old path keeps a re-export shim marked `# TEMPORARY: deleted by the
W2 sweep`. The `core/trackers` shim keeps its three backend adapters behind
`__getattr__`: eager re-export there would put mlflow/wandb/trackio in every
bare `import langres` and defeat tests/test_import_budget.py.

Callers and docs repointed in the follow-up commits.
